+++ app/api/budget/planned-income/route.ts
@@ -19,22 +19,6 @@ export async function POST (req: NextRequest) {
     return NextResponse.json({success: true}, {status: 201})
 }
 
-export async function DELETE (req: NextRequest) {
-    const userId = await getUserSessionId(req);
-    if (userId instanceof NextResponse) {
-        return userId;
-    }
-    
-    const requestBody = await req.json();
-    const errors = validateDELETEFields(requestBody)
-    if (errors.length) {
-        return NextResponse.json({success: false, error: `The following fields are missing from the request body: ${errors.join(", ")}` }, {status: 400});
-    }
-
-    await removePlannedIncome(new mongoose.Types.ObjectId(userId), requestBody.monthIndex, new mongoose.Types.ObjectId(requestBody.incomeSourceId));
-    return NextResponse.json({success: true}, {status: 200})
-}
-
 function validatePOSTFields (body: any): String[] {
     const missingFields: String[] = [];
     if(!body.monthIndex) {
@@ -42,17 +26,6 @@ function validatePOSTFields (body: any): String[] {
     }
     if(!body.newIncomeSource) {
         missingFields.push("newIncomeSource")
-    }
-    return missingFields;
-}
-
-function validateDELETEFields (body: any): String[] {
-    const missingFields: String[] = [];
-    if(!body.monthIndex) {
-        missingFields.push("monthIndex")
-    }
-    if(!body.incomeSourceId) {
-        missingFields.push("incomeSourceId")
     }
     return missingFields;
 }